revert: "Actively support py313: remove requests (#145)"
This reverts commit f363af092b3ef15e7c2491bd0446211605af747b.

Why: Seems there's some issue with SSE in our end2endtests which I've
also been able to reproduce locally though not consistently.
I tried in another branch to switch to niquest for everything (as
niquest supports SSE natively, we can get rid of sseclient-py library) but that
did not solve the issue.
Reverting that for now so I can move on with releasing other things